Add run makefile target
The target runs the most recently built "ship-it "docker container with
AWS env credentials. It's intended to be used with aws-vault, ie.
`aws-vault exec wattpad -- make run`

In the future, this will likely be replaced by kubernetes deployment
manifests (helm chart), but it's useful for early development in the
meanwhile.