Disable errcheck linter
The errcheck linter is a bit overzealous in that there's often cases
where it makes sense not to handle an error (for example, there's not
really anything sensible to do if logging an error fails. What would you
do, try to log the logging error?).

Of course we should still handle errors where appropriate, but we'll
have to be diligent about it ourselves since the tool has too many false
positives.