Add golint to travis GoLint build job (#20)
* Add golint to travis GoLint build job

* Fail the GoLint job when go lint complains

Otherwise, lint warnings won't cause the build to fail.

* Add golint shell script

golint exits successfully even when there are lint warnings, making it
awkward to integrate directly into CI. This small shell script allows us
to fail the CI build job if there are lint warnings and show the
warnings in the build log.

* Quell golint warnings about missing comments

* Replace golint with golangci-lint

The devs of golint are suprisingly adamant that golint should not be
used as part of CI. As an example: https://github.com/golang/lint/pull/148

The golangci-lint tool, on the other hand, is explicitly meant to be
used as part of CI and has better default behaviour for this use case.

* Use only golangci-lint for GoLint build job

golangci-lint is a meta linter that aggregates a number of other linting
tools. It includes staticcheck and go vet, so we don't need to run them
explicitly.

* Disable errcheck linter

The errcheck linter is a bit overzealous in that there's often cases
where it makes sense not to handle an error (for example, there's not
really anything sensible to do if logging an error fails. What would you
do, try to log the logging error?).

Of course we should still handle errors where appropriate, but we'll
have to be diligent about it ourselves since the tool has too many false
positives.