Velo 1457: helmrelease lister integration (#37)
* Add generated k8s code

* Move api types to the public `pkg` dir.
* Adds generated deepcopy, clientset and lister code for the `helmreleases.wattpad.com` api.

* Update go mod & vendored dependencies

* fixup! Add generated k8s code

* Add k8s-code-gen makefile target

This target runs the k8s code-generator tools to generate deepcopy,
lister, and clientset implementations for the custom api types.

In order to trick the generator tools into thinking they're running in a
GOPATH, the script fakes a GOPATH and symlinks the project repo to the
expected place.

* go mod vendor & tidy

* Update module dependencies

* update vendored deps

* Add build constraint to tool requires

We can't actually build this file, since `k8s.io/code-generator` isn't
an actual library. Adding an unsatisfied build constraint excludes the
file from the build.

* update generated deepcopy code

* Move tools import stub out of top-level

Moves the imports stub file into the dir of the tool that uses it.

* Add generated helmreleases informer code

* Update vendor dependencies

* Use helmreleases client in API k8s integration

Use the generated helmreleases API client to query for HelmRelease CRs
instead of ConfigMaps.

* Move struct definition and constructor func above methods

* polish k8s integration

* Use narrower `HelmReleaseGetter` interface
* Pre-allocate slice with known size

* Use informer/lister in k8s integration rather than direct clientset

The `HelmReleaseLister` interface provides caching as well as a slightly
nicer API for listing HelmRelease objects.

* Start HelmReleases informer with main context lifetime

The informer factory needs to be started before the informers it creates
can be used. It will be stopped when the service's main context is done.

* Make helmreleases resync interval configurable

defaults to 30s