Add gitsync/github pkg
This pkg implements the github event consumer functionality of the
gitsync component. It consumes push events from Github API which signal
that a chart has been changed, then syncs a helm release with the new
chart.

The consumer will be used to keep the ship-it-registry chart
automatically synced between github and k8s.

VELO-1452