Fix Infinite Loop in ECR Queue Consumer (#84)
* Fix Infinite Loop in ECR Queue Consumer

VELO-1529

The consumer returns an error when the image tag is not a valid commit
hash (ie. "latest"). The problem with this behaviour is that the SQS
consumer library believes the message has not been successfully
processed because the handler returned an error. As a result, the
invalid image is put back on the queue processed over and over in an
endless loop. This PR has the handler return nil on these types of
images so they are removed from the queue and logs the error instead.

* Make logs more actionable

VELO-1529

Implement PR Feedback