Fix Tiller Connection Issue (#87)
VELO-1529

When testing Ship-it in scal-labs, I discovered that that new releases
could not be installed due to a timing out connection to Tiller. The
default Tiller host used by the helm client is not the same as ours. This
PR adds an environment variable for the Tiller host and adds the value
as an option when creating the helm client.