Fixup syncd/github chart downloader (#92)
The downloader needs to preserve the full path of the file. Before it
would only save the filename, so a file like `foo/bar/file.yaml` would
be given a filename `file.yaml`. However when downloading a chart, which
contains nested directories like `chart/templates/*.yaml`, we need to
preserve the full file path in order to load the chart correctly.