VELO-1567: syncd batch image changes in a single commit (#114)
* Add bulk commit functionality to ECR reconciler

* Add 'image' internal package

This package provides the 'image.Ref' type: a reference to a remote
container image

* Add unstructured.VisitOne func

This func is used to traverse and mutate unstructured yaml data
represented as `yaml.MapSlice`. It's used to make order-preserving edits
to yaml files.

* Add ChartEditor implementation

The ChartEditor provides a method to update HelmRelease yaml files,
belonging to a specified Helm chart in Github, to use a desired image.
It batches the changes for a single image into one commit.

* Update syncd to use the new ChartEditor and image types

* Fixup golangci-lint Travis job

Turn off modules to workaround this issue:
https://github.com/golangci/golangci-lint/issues/730

increase the deadline from 1m -> 5m, we were seeing "deadline exceeded"
errors causing build failures

* go mod tidy

* Remove unused github env vars in ship-it-api config

* Remove unused github secrets from ship-it-api

The github secrets are not needed by the api, only syncd. For now, it's
assuming the secret is managed outside the chart, hence `existingSecretName`