Add watch & target namespace arguments for operator
The operator needs to know which namespace to watch for HelmRelease
resources (this will be the namespace where it is deployed), and which
namespace to tell Tiller to deploy releases into (this will probably be
"default").

This means ship-it can only deploy to one target namespace, in the
future maybe we'd like HelmReleases to be able to choose which namespace
they're deployed into.